// src/features/reports/report.types.ts
export type ReportRepairOrderTransaction = {
  id: string;

  roNumber: number;

  createdAt: string;

  customerId: string;

  customerName: string;

  vehicle: string;

  vin: string | null;

  technicianMembershipId: string | null;

  technicianName: string | null;

  serviceAdvisorMembershipId: string | null;

  serviceAdvisorName: string | null;

  laborHours: number;

  laborRevenue: number;

  partsRevenue: number;

  shopSupplies: number;

  discount: number;

  tax: number;

  total: number;

  status: string;

  priority: string;
};

//************************************************************** */

export type ReportFilterOption = {
  id: string;

  name: string;
};